+++ cypress/integration/professions/show.spec.ts
@@ -80,23 +80,16 @@ describe('Showing a profession', () => {
       'professions.show.qualification.level',
       'DSE - Diploma (post-secondary education), including Annex II (ex 92/51, Annex C,D) , Art. 11 c',
     );
-    cy.translate(
-      'professions.methodsToObtainQualification.generalSecondaryEducation',
-    ).then((method) => {
-      cy.checkSummaryListRowValue(
-        'professions.show.qualification.methods',
-        method,
-      );
-    });
 
-    cy.translate(
-      'professions.methodsToObtainQualification.generalSecondaryEducation',
-    ).then((method) => {
-      cy.checkSummaryListRowValue(
-        'professions.show.qualification.commonPath',
-        method,
-      );
-    });
+    cy.checkSummaryListRowValue(
+      'professions.show.qualification.routesToObtain',
+      'General secondary education',
+    );
+
+    cy.checkSummaryListRowValue(
+      'professions.show.qualification.mostCommonRouteToObtain',
+      'General secondary education',
+    );
 
     cy.checkSummaryListRowValue(
       'professions.show.qualification.duration',
@@ -109,6 +102,11 @@ describe('Showing a profession', () => {
         value,
       );
     });
+
+    cy.checkSummaryListRowValue(
+      'professions.show.qualification.moreInformationUrl',
+      'http://www.example.com',
+    );
 
     cy.translate('professions.show.legislation.heading').then((heading) => {
       cy.get('body').should('contain', heading);