+++ src/professions/admin/interfaces/qualifications.template.ts
@@ -1,11 +1,12 @@
 import { RadioButtonArgs } from '../../../common/interfaces/radio-button-args.interface';
 
-export interface QualificationInformationTemplate {
+export interface QualificationsTemplate {
   level: string;
-  methodToObtainQualificationRadioButtonArgs: RadioButtonArgs[];
-  mostCommonPathToObtainQualificationRadioButtonArgs: RadioButtonArgs[];
+  routesToObtain: string;
+  mostCommonRouteToObtain: string;
   duration: string;
   mandatoryProfessionalExperienceRadioButtonArgs: RadioButtonArgs[];
+  moreInformationUrl: string;
   ukRecognition: string;
   ukRecognitionUrl: string;
   otherCountriesRecognition: string;