+++ src/professions/admin/qualifications.controller.ts
@@ -17,10 +17,9 @@ import { Qualification } from '../../qualifications/qualification.entity';
 import { UserPermission } from '../../users/user-permission';
 import { ValidationFailedError } from '../../common/validation/validation-failed.error';
 import { ProfessionsService } from '../professions.service';
-import { MethodToObtainQualificationRadioButtonsPresenter } from './method-to-obtain-qualification-radio-buttons.presenter';
 import { YesNoRadioButtonArgsPresenter } from './yes-no-radio-buttons-presenter';
-import { QualificationInformationDto } from './dto/qualification-information.dto';
-import { QualificationInformationTemplate } from './interfaces/qualification-information.template';
+import { QualificationsDto } from './dto/qualifications.dto';
+import { QualificationsTemplate } from './interfaces/qualifications.template';
 import { BackLink } from '../../common/decorators/back-link.decorator';
 import ViewUtils from './viewUtils';
 import { ProfessionVersionsService } from '../profession-versions.service';
@@ -30,14 +29,14 @@ import { ProfessionVersion } from '../profession-version.entity';
 
 @UseGuards(AuthenticationGuard)
 @Controller('admin/professions')
-export class QualificationInformationController {
+export class QualificationsController {
   constructor(
     private readonly professionsService: ProfessionsService,
     private readonly professionVersionsService: ProfessionVersionsService,
     private readonly i18nService: I18nService,
   ) {}
 
-  @Get('/:professionId/versions/:versionId/qualification-information/edit')
+  @Get('/:professionId/versions/:versionId/qualifications/edit')
   @Permissions(UserPermission.CreateProfession, UserPermission.EditProfession)
   @BackLink((request: Request) =>
     request.query.change === 'true'
@@ -67,7 +66,7 @@ export class QualificationInformationController {
     );
   }
 
-  @Post('/:professionId/versions/:versionId/qualification-information')
+  @Post('/:professionId/versions/:versionId/qualifications')
   @Permissions(UserPermission.CreateProfession, UserPermission.EditProfession)
   @BackLink((request: Request) =>
     request.body.change === 'true'
@@ -78,11 +77,11 @@ export class QualificationInformationController {
     @Res() res: Response,
     @Param('professionId') professionId: string,
     @Param('versionId') versionId: string,
-    @Body() qualificationInformationDto,
+    @Body() qualificationsDto,
   ): Promise<void> {
     const validator = await Validator.validate(
-      QualificationInformationDto,
-      qualificationInformationDto,
+      QualificationsDto,
+      qualificationsDto,
     );
 
     const profession = await this.professionsService.findWithVersions(
@@ -93,8 +92,7 @@ export class QualificationInformationController {
       versionId,
     );
 
-    const submittedValues: QualificationInformationDto =
-      qualificationInformationDto;
+    const submittedValues: QualificationsDto = qualificationsDto;
 
     const mandatoryProfessionalExperience =
       submittedValues.mandatoryProfessionalExperience === undefined
@@ -105,13 +103,11 @@ export class QualificationInformationController {
       ...version.qualification,
       ...{
         level: submittedValues.level,
-        methodToObtain: submittedValues.methodToObtainQualification,
-        otherMethodToObtain: submittedValues.otherMethodToObtainQualification,
-        commonPathToObtain: submittedValues.mostCommonPathToObtainQualification,
-        otherCommonPathToObtain:
-          submittedValues.otherMostCommonPathToObtainQualification,
+        routesToObtain: submittedValues.routesToObtain,
+        mostCommonRouteToObtain: submittedValues.mostCommonRouteToObtain,
         educationDuration: submittedValues.duration,
         mandatoryProfessionalExperience,
+        url: submittedValues.moreInformationUrl,
         ukRecognition: submittedValues.ukRecognition,
         ukRecognitionUrl: submittedValues.ukRecognitionUrl,
         otherCountriesRecognition: submittedValues.otherCountriesRecognition,
@@ -156,33 +152,18 @@ export class QualificationInformationController {
     change: boolean,
     errors: object | undefined = undefined,
   ) {
-    const mostCommonPathToObtainQualificationRadioButtonArgs =
-      await new MethodToObtainQualificationRadioButtonsPresenter(
-        qualification?.commonPathToObtain,
-        qualification?.otherCommonPathToObtain,
-        errors,
-        this.i18nService,
-      ).radioButtonArgs('mostCommonPathToObtainQualification');
-
-    const methodToObtainQualificationRadioButtonArgs =
-      await new MethodToObtainQualificationRadioButtonsPresenter(
-        qualification?.methodToObtain,
-        qualification?.otherMethodToObtain,
-        errors,
-        this.i18nService,
-      ).radioButtonArgs('methodToObtainQualification');
-
     const mandatoryProfessionalExperienceRadioButtonArgs =
       await new YesNoRadioButtonArgsPresenter(
         qualification?.mandatoryProfessionalExperience,
         this.i18nService,
       ).radioButtonArgs();
 
-    const templateArgs: QualificationInformationTemplate = {
+    const templateArgs: QualificationsTemplate = {
       level: qualification?.level,
-      methodToObtainQualificationRadioButtonArgs,
-      mostCommonPathToObtainQualificationRadioButtonArgs,
+      routesToObtain: qualification?.routesToObtain,
+      mostCommonRouteToObtain: qualification?.mostCommonRouteToObtain,
       mandatoryProfessionalExperienceRadioButtonArgs,
+      moreInformationUrl: qualification?.url,
       duration: qualification?.educationDuration,
       captionText: ViewUtils.captionText(isEditing),
       ukRecognition: qualification?.ukRecognition,
@@ -194,9 +175,6 @@ export class QualificationInformationController {
       errors,
     };
 
-    return res.render(
-      'admin/professions/qualification-information',
-      templateArgs,
-    );
+    return res.render('admin/professions/qualifications', templateArgs);
   }
 }