+++ src/professions/interfaces/show-template.interface.ts
@@ -1,12 +1,12 @@
+import { SummaryList } from '../../common/interfaces/summary-list';
 import { Organisation } from '../../organisations/organisation.entity';
-import QualificationPresenter from '../../qualifications/presenters/qualification.presenter';
 import { ProfessionPresenter } from '../presenters/profession.presenter';
 import { Profession } from '../profession.entity';
 
 export interface ShowTemplate {
   profession: Profession;
   presenter?: ProfessionPresenter;
-  qualification: QualificationPresenter;
+  qualificationSummaryList: SummaryList;
   nations: string[];
   industries: string[];
   organisation: Organisation;