+++ src/professions/professions.controller.ts
@@ -51,12 +51,14 @@ export class ProfessionsController {
     );
 
     const qualification = profession.qualification
-      ? new QualificationPresenter(profession.qualification)
+      ? new QualificationPresenter(profession.qualification, this.i18nService)
       : null;
 
     return {
       profession,
-      qualification: qualification,
+      qualificationSummaryList: qualification
+        ? await qualification.summaryList()
+        : null,
       nations,
       industries,
       organisation,