+++ src/qualifications/presenters/qualification.presenter.spec.ts
@@ -1,81 +1,50 @@
-import { MethodToObtain } from '../qualification.entity';
 import qualificationFactory from '../../testutils/factories/qualification';
 import QualificationPresenter from './qualification.presenter';
 import { formatMultilineString } from '../../helpers/format-multiline-string.helper';
 import { multilineOf } from '../../testutils/multiline-of';
+import { createMockI18nService } from '../../testutils/create-mock-i18n-service';
+import { translationOf } from '../../testutils/translation-of';
 
 jest.mock('../../helpers/format-multiline-string.helper');
 
 describe(QualificationPresenter, () => {
-  describe('methodToObtainQualification', () => {
-    describe('when the method to ObtainQualification is "others"', () => {
-      it('returns the other text value', () => {
-        (formatMultilineString as jest.Mock).mockImplementation(multilineOf);
+  describe('routesToObtain', () => {
+    it('returns the a multiline string of the text value', () => {
+      (formatMultilineString as jest.Mock).mockImplementation(multilineOf);
 
-        const qualification = qualificationFactory.build({
-          methodToObtain: MethodToObtain.Others,
-          otherMethodToObtain: 'other value',
-        });
-
-        const presenter = new QualificationPresenter(qualification);
-
-        expect(presenter.methodToObtainQualification).toEqual(
-          multilineOf('other value'),
-        );
-
-        expect(formatMultilineString).toBeCalledWith('other value');
+      const qualification = qualificationFactory.build({
+        routesToObtain: 'other value',
       });
-    });
 
-    describe('when the method to ObtainQualification is not "others"', () => {
-      it('returns the localisation id for the selected method', () => {
-        const qualification = qualificationFactory.build({
-          methodToObtain: MethodToObtain.DegreeLevel,
-          otherMethodToObtain: 'other value',
-        });
+      const presenter = new QualificationPresenter(
+        qualification,
+        createMockI18nService(),
+      );
 
-        const presenter = new QualificationPresenter(qualification);
+      expect(presenter.routesToObtain).toEqual(multilineOf('other value'));
 
-        expect(presenter.methodToObtainQualification).toEqual(
-          'professions.methodsToObtainQualification.degreeLevel',
-        );
-      });
+      expect(formatMultilineString).toBeCalledWith('other value');
     });
   });
 
-  describe('mostCommonPathToObtainQualification', () => {
-    describe('when the method to ObtainQualification is "others"', () => {
-      it('returns the other text value', () => {
-        (formatMultilineString as jest.Mock).mockImplementation(multilineOf);
+  describe('mostCommonRouteToObtain', () => {
+    it('returns the a multiline string of the text value', () => {
+      (formatMultilineString as jest.Mock).mockImplementation(multilineOf);
 
-        const qualification = qualificationFactory.build({
-          commonPathToObtain: MethodToObtain.Others,
-          otherCommonPathToObtain: 'other value',
-        });
-
-        const presenter = new QualificationPresenter(qualification);
-
-        expect(presenter.mostCommonPathToObtainQualification).toEqual(
-          multilineOf('other value'),
-        );
-
-        expect(formatMultilineString).toBeCalledWith('other value');
+      const qualification = qualificationFactory.build({
+        mostCommonRouteToObtain: 'other value',
       });
-    });
 
-    describe('when the method to ObtainQualification is not "others"', () => {
-      it('returns the localisation id for the selected method', () => {
-        const qualification = qualificationFactory.build({
-          commonPathToObtain: MethodToObtain.DegreeLevel,
-          otherCommonPathToObtain: 'other value',
-        });
+      const presenter = new QualificationPresenter(
+        qualification,
+        createMockI18nService(),
+      );
 
-        const presenter = new QualificationPresenter(qualification);
+      expect(presenter.mostCommonRouteToObtain).toEqual(
+        multilineOf('other value'),
+      );
 
-        expect(presenter.mostCommonPathToObtainQualification).toEqual(
-          'professions.methodsToObtainQualification.degreeLevel',
-        );
-      });
+      expect(formatMultilineString).toBeCalledWith('other value');
     });
   });
 
@@ -86,7 +55,10 @@ describe(QualificationPresenter, () => {
           mandatoryProfessionalExperience: true,
         });
 
-        const presenter = new QualificationPresenter(qualification);
+        const presenter = new QualificationPresenter(
+          qualification,
+          createMockI18nService(),
+        );
 
         expect(presenter.mandatoryProfessionalExperience).toEqual('app.yes');
       });
@@ -98,9 +70,45 @@ describe(QualificationPresenter, () => {
           mandatoryProfessionalExperience: false,
         });
 
-        const presenter = new QualificationPresenter(qualification);
+        const presenter = new QualificationPresenter(
+          qualification,
+          createMockI18nService(),
+        );
 
         expect(presenter.mandatoryProfessionalExperience).toEqual('app.no');
+      });
+    });
+  });
+
+  describe('moreInformationUrl', () => {
+    describe('when a blank string is provided', () => {
+      it('returns null', () => {
+        const qualification = qualificationFactory.build({
+          url: '',
+        });
+
+        const presenter = new QualificationPresenter(
+          qualification,
+          createMockI18nService(),
+        );
+
+        expect(presenter.moreInformationUrl).toEqual(null);
+      });
+    });
+    describe('when a URL is provided', () => {
+      it('returns a link', () => {
+        const qualification = qualificationFactory.build({
+          url: 'http://example.com',
+        });
+
+        const presenter = new QualificationPresenter(
+          qualification,
+          createMockI18nService(),
+        );
+
+        expect(presenter.moreInformationUrl).toEqual(
+          '<a class="govuk-link" href="http://example.com">http://example.com</a>',
+        );
       });
     });
   });
@@ -112,7 +120,10 @@ describe(QualificationPresenter, () => {
           ukRecognitionUrl: '',
         });
 
-        const presenter = new QualificationPresenter(qualification);
+        const presenter = new QualificationPresenter(
+          qualification,
+          createMockI18nService(),
+        );
 
         expect(presenter.ukRecognitionUrl).toEqual(null);
       });
@@ -123,7 +134,10 @@ describe(QualificationPresenter, () => {
           ukRecognitionUrl: 'http://example.com',
         });
 
-        const presenter = new QualificationPresenter(qualification);
+        const presenter = new QualificationPresenter(
+          qualification,
+          createMockI18nService(),
+        );
 
         expect(presenter.ukRecognitionUrl).toEqual(
           '<a class="govuk-link" href="http://example.com">http://example.com</a>',
@@ -139,7 +153,10 @@ describe(QualificationPresenter, () => {
           otherCountriesRecognitionUrl: '',
         });
 
-        const presenter = new QualificationPresenter(qualification);
+        const presenter = new QualificationPresenter(
+          qualification,
+          createMockI18nService(),
+        );
 
         expect(presenter.otherCountriesRecognitionUrl).toEqual(null);
       });
@@ -150,11 +167,89 @@ describe(QualificationPresenter, () => {
           otherCountriesRecognitionUrl: 'http://example.com',
         });
 
-        const presenter = new QualificationPresenter(qualification);
+        const presenter = new QualificationPresenter(
+          qualification,
+          createMockI18nService(),
+        );
 
         expect(presenter.otherCountriesRecognitionUrl).toEqual(
           '<a class="govuk-link" href="http://example.com">http://example.com</a>',
         );
+      });
+    });
+  });
+
+  describe('summaryList', () => {
+    it('returns a summary list of all Qualification fields', async () => {
+      const qualification = qualificationFactory.build({
+        otherCountriesRecognitionUrl: 'http://example.com',
+      });
+
+      const presenter = new QualificationPresenter(
+        qualification,
+        createMockI18nService(),
+      );
+
+      expect(presenter.summaryList()).resolves.toEqual({
+        classes: 'govuk-summary-list--no-border',
+        rows: [
+          {
+            key: {
+              text: translationOf('professions.show.qualification.level'),
+            },
+            value: {
+              html: formatMultilineString(presenter.level),
+            },
+          },
+          {
+            key: {
+              text: translationOf(
+                'professions.show.qualification.routesToObtain',
+              ),
+            },
+            value: {
+              html: presenter.routesToObtain,
+            },
+          },
+          {
+            key: {
+              text: translationOf(
+                'professions.show.qualification.mostCommonRouteToObtain',
+              ),
+            },
+            value: {
+              html: presenter.mostCommonRouteToObtain,
+            },
+          },
+          {
+            key: {
+              text: translationOf('professions.show.qualification.duration'),
+            },
+            value: {
+              text: presenter.duration,
+            },
+          },
+          {
+            key: {
+              text: translationOf(
+                'professions.show.qualification.mandatoryExperience',
+              ),
+            },
+            value: {
+              text: translationOf(presenter.mandatoryProfessionalExperience),
+            },
+          },
+          {
+            key: {
+              text: translationOf(
+                'professions.show.qualification.moreInformationUrl',
+              ),
+            },
+            value: {
+              html: presenter.moreInformationUrl,
+            },
+          },
+        ],
       });
     });
   });