+++ CHANGELOG.md
@@ -8,7 +8,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ### Added
 
+- Add "more information url" to Qualification detail entry page
+
 ### Changed
+
+- Replace routes to obtain Qualification & most common path to obtain radio buttons with textareas
+- Remove Qualification Level from admin organisation list
+- Remove deprecated fields from Qualification
 
 ## [release-004] - 2022-02-21